--- routes/article.route.js
@@ -1,9 +1,0 @@
-const express = require("express")
-const router = express.Router()
-
-const ArticleController = require("../controllers/article.controller")
-const articleController = new ArticleController()
-
-router.get("/", articleController.findAllArticles)
-
-module.exports = router
